portfolio provisioning. Move Azure stages into a mixin module. Logic in triggering next transition moved out of Celery task into state machine class
